modules/Controller/requirements.txt
asyncio==3.4.3
azure-iot-device==2.4.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
deprecation==2.1.0
idna==2.10
janus==0.4.0
packaging==20.8
paho-mqtt==1.5.1
pigpio==1.78
pyparsing==2.4.7
PySocks==1.7.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
requests-unixsocket==0.2.0
six==1.15.0
urllib3==1.25.11
w1thermsensor==1.3.0
